Fixed bug, improved NameComparator class.
There was a bug in the part of the code designed to throw out results that
matched on legal names when preferred names did not. This was due to the lack
of caseless comparison when checking to make sure the one string contained
the other and has been taken care of.

In case of future use, the NameComparator class no longer has extractSecondaryName()
and extractTertiaryName() methods and now simply uses an extractTiebreakerNames()
method.